:truck: Move assessments landing page into a folder
