+++ libs/features/convs-mgr/conversations/assessments/src/index.ts
@@ -1,4 +1,4 @@
 export * from './lib/convs-mgr-assessments.module';
 export * from './lib/services/assessment-metrics.service';
 export * from './lib/components/content-section/content-section.component';
-export * from './lib/components/assessment-landing-page.component';
+export * from './lib/components/micro-app-assessment-landing-page/assessment-landing-page.component';
